requirements.txt
flask
groq
flask-cors
gunicorn

python-dotenv
